Authorize answer

+++ app/Http/Controllers/AnswerController.php
@@ -6,6 +6,7 @@ use App\Http\Requests\StoreAnswerRequest;
 use App\Http\Requests\UpdateAnswerRequest;
 use App\Models\Answer;
 use App\Models\Question;
+use Illuminate\Support\Facades\Gate;
 
 class AnswerController extends Controller
 {
@@ -60,6 +61,8 @@ class AnswerController extends Controller
      */
     public function update(UpdateAnswerRequest $request, Question $question, Answer $answer)
     {
+        // Gate::authorize('update', $answer);
+
         $answer->update($request->validated());
 
         return back()->with('success', 'Your answer updated successfully.');
@@ -70,6 +73,8 @@ class AnswerController extends Controller
      */
     public function destroy(Question $question, Answer $answer)
     {
+        Gate::authorize('delete', $answer);
+
         $answer->delete();
 
         return back()->with('success', 'Your answer deleted successfully.');
